MOved the  rag for the llms intro the llm interface rather than in the specific lmstudio instance.
